requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2019.11.28
cffi==1.13.2
decorator==4.4.1
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
joblib==0.14.1
jupyter-client==5.3.4
jupyter-core==4.6.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
munch==2.5.0
numpy==1.18.1
opencv-python==4.1.2.30
pandas==0.25.3
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
pretrainedmodels==0.7.4
prompt-toolkit==3.0.2
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.5.2
python-dateutil==2.8.1
pytz==2019.3
pyzmq==18.1.1
scikit-learn==0.22.1
scipy==1.3.1
six==1.13.0
torch==1.4.0
torchvision==0.5.0
tornado==6.0.3
tqdm==4.41.1
traitlets==4.3.3
wcwidth==0.1.8